+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 1.1.4 - 2025.08.01
+
+### Added
+
+### Changed
+
+- Replace `:` with `_` when creating host directories in `output/`
+- Rename `esquery` engine to `ast` engine
+- Allow any string type for `CheckAssignmentExistStep`
+
+### Fixed
+
+- Fix the issue in ast engine to handle direct assignments in `checkAssignmentExist`
+
 ## 1.1.4-alpha.4 - 2025.07.31
 
 ### Added